Added script position switching
Added buttons from moving scripts up and down in a subaction. Currently,
once a script is moved the subaction is automatically saved. That is
subject to change. There's some unoptimized/messy code at the moment,
but from the testing I've done it works properly.

The GUI for it is a little ugly at the moment, but I'll polish
everything once I have the actual code done.

I also added a small feature for convenience: Crazy Hand will
automatically go to the directory of the ISO it loaded during its last
session.