Special moves 2: Electric boogaloo
Crammed special moves in; There might be crashes with master/crazy hand
and wireframes, but I just wanted to push this out there so we all have
it to play around with. If anyone wants to help map out the names for
custom moves at some point that'd be great, but it's not necessary.

Next on the to-do list is more script options.