First draft of script switching
Added a drop-down menu on every script that allows you to change the
type of script it is. At the moment, some issues can arise from
switching to a script of a different size, as is outlined in Script. In
addition, a small change has been made to the "set loop" event; It turns
out the size of it is 0x8, not 0x4. C0 00 00 XX sets how many times(X)
the loop should be played, while the next four entries(04 00 00 YY)
decides how many frames to wait(Y) before executing the loop again.
Thanks to Itaru on smashboards for that one :)

Plans for the near future:
Polish script switching to prevent the current flaws that were
previously mentioned
Make a "loop script" for ease of use.

I'm going to try and figure out EXACTLY how projectiles are spawned.
Being able to insert a script for spawning projectiles would be handy.