Loop script bugfix, and a little bit more...
Fixed a bug with loop scripts where loops would not properly execute. In
addition, I was wrong about how they worked(Asynchronous timers are
needed in the script to specify how often to do the loop, but they are
not always directly after the set loop function).

In addition, I added a visual cue to show the user if a script is inside
a loop. It's pretty ugly at the moment, but functionality is my main
concern for dev builds.

I also added a button to open an ISO directly in dolphin(Currently only
for windows). I still need to polish a couple parts of it, but it is
functional and fairly convenient :)