Preliminary file menu, outline for help menu.
-Added a file menu that has options to open another ISO, and another
option to close the program. More features will be added to the menu(A
button for opening the ISO in dolphin for instance, once I write the
code for opening ISO's in other operating systems.) In the long run I'd
like to make a good amount of the utility options we've been putting in
RestorePanel into options on the menu bar for quicker access.