Tater's big book of changes
-Cleaned up script switching code
-Added "special attributes" menu; not all characters are mapped yet.
-Added saving/loading characters as .DAT files via File menu
-Working on a format for saving/loading subactions(see below)
-Added Script editor window(Incomplete!)
-Changed how scripts work a little(see below)

It's been a busy week for me, and it's going to be busy for a little
while longer. However, I've still managed to squeeze in time to program.
I'm committing the changes I've made now because I've changed how
scripts work somewhat and don't want to cause much inconvenience for
anyone else.

For now, Script.scripts will be used for whatever subaction is loaded
inside CrazyHand as it always has been. However, because scripts were
originally hardcoded to be in Script.scripts, and save themselves to the
current character file whenever anything was changed, it was impossible
to implement what I had in mind; A separate script editing window in
which the user could open multiple subactions at a time and freely
edit/save without the hassle of switching from character to character
each time. Each subaction can be saved or loaded as a .subact(or a
prettier file extension, idk) file. This will address my own concerns
with messing up a subaction by putting the wrong script in the wrong
place, as well as just being a good feature to have. In addition, there
will be the option to create a tab that isn't linked to any character
file whatsoever. Sort of a "sandbox mode" for scripting. At the moment,
I don't know of any way to inject subaction data into the free space
within the ISO, but if we do find a way this feature would allow users
to potentially create whole new moves.

Anyways, due to scripts trying to save themselves no matter what window
they're in I had to rewrite a chunk of how they work.

setStandAlone() makes the script part of a "sandbox subaction". Not much
more to say about that one.

setLinkedToCharacterSubaction(Character, offset) will be used to link
scripts to a character and the offset for the subaction they are located
in. I think I'll just move this to ScriptEditWindowTab, since that would
be easier. Ignore this one haha

SetInEditor(Editor) tells the script it's inside the script editor
window, not the Crazy Hand window.

toString() is being used for the data structure of .subact files. Not
quite finished yet.

ScriptUtils contains all the utility methods for scripts, both that I've
made and that were lying around in MeleeEdit/Script.

Methods in FileIO are still somewhat of a mess; Ignore them as they'll
be gone soon(dumpCharacterData was an interesting experiment... Didn't
pan out.)