More progress on FSM
Made a bit more progress on it today.
You can now save the FSM's, however you currently cannot add new ones.
So if you open a file that already has modifiers you can change them but
that's currently it.
I'll probably get a chance to work on it more tomorrow.