Removed randomization due to bug
Randomization apparently produces a fatal error now, and I can't seem to
find a fix for it. For the time being, I'm just going to comment out the
option for randomization for the sake of the next release. I figure that
it's not an important feature to lose sleep over.