A couple bugfixes, added random seed feature
-Fixed saving/loading characters into the ISO not working fully as
intended(Turns out it I just forgot to tell CrazyHand to close the ISO
file, so when I tried to run dolphin after loading a character, Dolphin
thought the iso was corrupt because it was in use by CrazyHand).

-Fixed garbage data problems when replacing a script with a new script
of a different size(I.E making a timer into a hitbox).

-Fixed randomization error caused by my own changes to FileIO.

Added feature to set the seed for character randomization. In addition,
when a seed is not used for the randomizer CrazyHand will make a small
.txt file named randomseeds that contains all the seeds that the
randomizer used when a seed was not given, and what date/time it was
created.