Bugfixes, Introduction of DOLPatch.
-Fixed bug with newly created Sound Effect scripts not behaving
properly.
-Fixed bug where "hitbox interaction" values were not saving.
-Fixed randomization not randomizing hitboxes.
-Optimized randomization speed some; Java was creating new instances of
every GUI node in a script every time we called a script to be
randomized, which caused it to gobble memory and finally crash. Scripts
now check if FileIO is randomizing character data to determine if they
should initialize GUI elements or not. The biggest change this makes is
in HitboxScript where each text field is now initialized after checking
to see if randomization is in-progress, whereas before they were
initialized as soon as the object was constructed. In addition, when
calling scramble() for scripts, we should refrain from using any GUI
elements for gathering the data for the same reason.

-Added randomization parameters for LoopScript and SynchronousScript.
Both will set their value somewhere between 0.5x to 1.5x their normal
values.

-Added DOLPatch class. There is (semi)detailed documentation in the
class itself. As I learn more about DOL injection, I hope to compile a
list of commonly desired codes and add them to this list. For the
moment, it is a proof-of-concept that fully functions. Also a pretty
nifty code :)

-Created SelfDamageScript, but left it unimplemented for now. Further
details are inside the class.

-Removed some random debug System.out.println()'s that are not needed.