Just keeping the Github project update to date with my work
This is by no means a final version, but there have been a good amount
of changes, so I felt it was time to post an update

Brief changelist:
-Added DOLPatch creation window. Pretty ugly right now, but it's
functional.
-Added DOLPatch for Achille's FSM engine! It currently injects FSM data
from SD remix, however... Need to fix that.
-Added ProjectileEditPane and some offsets for different projectile
hitboxes.
-Added static character ID fields for convenience(I.E DRMARIO_ID instead
of 2)
-Moved DOLPatch to a new package for better organization

There are probably some other minor changes here and there, but they
aren't important or are temporary.