Added projectile spawning support
--added support for Achilles projectile spawning event.
--added ItCo.dat to the list of editable files.
--started adding support for editing hitboxes for non-character
projectiles. It currently isn't functioning properly though, so I've
left it out of this release. The code is all working properly to load
and edit hitboxes from ItCo, but it seems that the hitboxes found in
ItCo aren't linked to the projectiles the way I think they are. For now
this stuff is commented out until I get a chance to do more research.
--changed the color of the self damage script to a more attractive color
:)