Progress so far
I wasn't planning to publish this on GitHub until I finished it due to
the fact that I've changed so many things and I've broken many things in
the process, but I want to be able to work from multiple computers so I
decided to upload what I have.

If anyone sees this, I strongly recommend you just steer away from this
branch until I'm done. I'm working on a pretty big revamp of CH to make
it more user friendly and this is what I have so far. I'll go into more
detail on what I've added in the future.